+++ src/Console/System/CheckStatusCommand.php
@@ -40,7 +40,6 @@ use Glpi\System\Status\StatusChecker;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Toolbox;
 
 class CheckStatusCommand extends AbstractCommand
 {
@@ -52,13 +51,6 @@ class CheckStatusCommand extends AbstractCommand
 
         $this->setName('system:status');
         $this->setDescription(__('Check system status'));
-        $this->addOption(
-            'format',
-            'f',
-            InputOption::VALUE_OPTIONAL,
-            'Output format [plain or json]',
-            'plain'
-        );
         $this->addOption(
             'private',
             'p',
@@ -76,16 +68,8 @@ class CheckStatusCommand extends AbstractCommand
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-
-        $format = strtolower($input->getOption('format'));
-        $status = StatusChecker::getServiceStatus($input->getOption('service'), !$input->getOption('private'), $format === 'json');
-
-        if ($format === 'json') {
-            $output->writeln(json_encode($status, JSON_PRETTY_PRINT));
-        } else {
-            Toolbox::deprecated('Plain-text status output is deprecated please use the JSON format instead by specifically using the "--format json" parameter. In the future, JSON output will be the default.');
-            $output->writeln($status);
-        }
+        $status = StatusChecker::getServiceStatus($input->getOption('service'), !$input->getOption('private'));
+        $output->writeln(json_encode($status, JSON_PRETTY_PRINT));
 
         return 0; // Success
     }